updata config menu

+++ Joypad_RC/mwc.h
@@ -4,60 +4,60 @@
 #include <ZigduinoRadio.h>
 #endif
 
-int RCin[8],RCoutA[8],RCoutB[8];
+int16_t RCin[8], RCoutA[8], RCoutB[8];
 
-int p;
-uint16_t read16() 
+int16_t p;
+uint16_t read16()
 {
-  uint16_t r = (inBuf[p++]&0xFF);
-  r+= (inBuf[p++]&0xFF)<<8;
+  uint16_t r = (inBuf[p++] & 0xFF);
+  r += (inBuf[p++] & 0xFF) << 8;
   return r;
 }
 
-uint16_t t,t1,t2;
-uint16_t write16(boolean a) 
+uint16_t t, t1, t2;
+uint16_t write16(boolean a)
 {
-  if(a)
+  if (a)
   {
-    t1=outBuf[p++]>>8;
-    t2=outBuf[p-1]-(t1<<8);
-    t=t1;
+    t1 = outBuf[p++] >> 8;
+    t2 = outBuf[p - 1] - (t1 << 8);
+    t = t1;
   }
   else
-    t=t2;
+    t = t2;
   return t;
 }
 
 typedef  unsigned char byte;
-byte getChecksum(byte length,byte cmd,byte mydata[])
+byte getChecksum(byte length, byte cmd, byte mydata[])
 {
   //三个参数分别为： 数据长度  ，  指令代码  ，  实际数据数组
-  byte checksum=0;
-  checksum ^= (length&0xFF);
-  checksum ^= (cmd&0xFF);
-  for(int i=0;i<length;i++)
+  byte checksum = 0;
+  checksum ^= (length & 0xFF);
+  checksum ^= (cmd & 0xFF);
+  for (uint8_t i = 0; i < length; i++)
   {
-    checksum ^= (mydata[i]&0xFF);
+    checksum ^= (mydata[i] & 0xFF);
   }
   return checksum;
-} 
+}
 
 void data_rx()
 {
-  //  s_struct_w((uint8_t*)&inBuf,16);
-  p=0;
-  for(int i=0;i<8;i++) 
+  //  s_struct_w((int*)&inBuf,16);
+  p = 0;
+  for (uint8_t i = 0; i < 8; i++)
   {
-    RCin[i]=read16();
+    RCin[i] = read16();
     /*
     Serial.print("RC[");
      Serial.print(i+1);
      Serial.print("]:");
-     
+
      Serial.print(inBuf[2*i],DEC);
      Serial.print(",");
      Serial.print(inBuf[2*i+1],DEC);
-     
+
      Serial.print("---");
      Serial.println(RCin[i]);
      */
@@ -67,25 +67,25 @@ void data_rx()
 
 void data_tx()
 {
-  p=0;
-  for(int i=0;i<8;i++) 
+  p = 0;
+  for (uint8_t i = 0; i < 8; i++)
   {
-    RCoutA[i]=write16(1);
-    RCoutB[i]=write16(0);
+    RCoutA[i] = write16(1);
+    RCoutB[i] = write16(0);
 
     /*
     Serial.print("RC[");
      Serial.print(i+1);
      Serial.print("]:");
-     
+
      Serial.print(RCout[i]);
-     
+
      Serial.print("---");
-     
+
      Serial.print(RCoutA[i],DEC);
      Serial.print(",");
      Serial.print(RCoutB[i],DEC);
-     
+
      Serial.println("");
      */
     //    delay(50);        // delay in between reads for stability
@@ -104,23 +104,23 @@ void data_send()
 
 #if !defined(__AVR_ATmega128RFA1__)
   static byte buf_head[3];
-  buf_head[0]=0x24;  
-  buf_head[1]=0x4D;  
-  buf_head[2]=0x3C;
+  buf_head[0] = 0x24;
+  buf_head[1] = 0x4D;
+  buf_head[2] = 0x3C;
 #endif
 
-  #define buf_length 0x10   //16
-  #define buf_code 0xC8     //200
-  
+#define buf_length 0x10   //16
+#define buf_code 0xC8     //200
+
   static byte buf_data[buf_length];
-  for(int a=0;a<(buf_length/2);a++)
+  for (uint8_t a = 0; a < (buf_length / 2); a++)
   {
-    buf_data[2*a]=RCoutB[a];
-    buf_data[2*a+1]=RCoutA[a];
+    buf_data[2 * a] = RCoutB[a];
+    buf_data[2 * a + 1] = RCoutA[a];
   }
 
   static byte buf_body;
-  buf_body=getChecksum(buf_length,buf_code,buf_data);
+  buf_body = getChecksum(buf_length, buf_code, buf_data);
 
   //----------------------
 #if defined(__AVR_ATmega128RFA1__)
@@ -129,13 +129,13 @@ void data_send()
   mwc_port.write(0xbb);
   mwc_port.write(0xcc);
 #else
-  for(int a=0;a<3;a++){
+  for (uint8_t a = 0; a < 3; a++) {
     mwc_port.write(buf_head[a]);
   }
   mwc_port.write(buf_length);
   mwc_port.write(buf_code);
 #endif
-  for(int a=0;a<buf_length;a++){
+  for (uint8_t a = 0; a < buf_length; a++) {
     mwc_port.write(buf_data[a]);
   }
   mwc_port.write(buf_body);
